kinesis-firehose-stream/requirements.txt
boto3==1.35.79
aws-cdk-lib==2.173.2
constructs>=10.0.0,<11.0.0
aws-cdk.aws-dynamodb
aws-cdk.aws-s3
aws-cdk.aws-kinesis
aws-cdk.aws-kinesisfirehose
aws-cdk.cx-api
